Current successful request and cheerio selectors for all vendor data
This example has been heavily modified for my needs and contains successful selectors and a successful export of the data using the fs.writeFile method and fs-extra module for one individual dealer. The next step was to start creating layers fucntions that take arrays and pass out arrays of data.