Multiple test here. Succesful interation through a multitude of links.
Created a callback function that iterated through a list of links and returned each one as a json like object and then put them in an array
Next in the callback I figured out how to export that JSON data into a file using fs stream. This was faciltated by the module fs-extraand the use of JSON.stringify.  This was the final successful codeAcquires all the links to each INDIVIDUAL STATE