Tried messing with the Q promise but not successfully
I really wanted to deconstruct this example to see what parts were relying on what other parts. I was trying grasp when to request the response and how exactly to do that. For some reason the data that is gleaned in the rp() wont go to vendor except the url. I just read that maybe I need another promise.all here that will grab the data after the urls have been visited.??